requirements.txt
flask
mafia_schedule
azure-storage-blob
python-decouple

